refactor(code): Clean up codes
