+++ androidApp/src/main/java/id/nisyafawwaz/nyampur/android/adapters/SavedRecipeAdapter.kt
@@ -9,6 +9,7 @@ import com.bumptech.glide.Glide
 import id.nisyafawwaz.nyampur.android.R
 import id.nisyafawwaz.nyampur.android.databinding.ItemSavedRecipeBinding
 import id.nisyafawwaz.nyampur.domain.models.RecipeModel
+import java.util.Locale
 
 class SavedRecipeAdapter : RecyclerView.Adapter<SavedRecipeAdapter.ViewHolder>() {
 
@@ -35,7 +36,7 @@ class SavedRecipeAdapter : RecyclerView.Adapter<SavedRecipeAdapter.ViewHolder>()
         fun bind(recipe: RecipeModel) {
             with(binding) {
                 tvFoodName.text = recipe.title
-                tvCookTime.text = recipe.cookTime.toString()
+                tvCookTime.text = String.format(Locale.getDefault(), "%d", recipe.cookTime)
                 tvLevel.text = recipe.level
 
                 Glide.with(root.context)
