+++ shared/src/androidMain/kotlin/id/nisyafawwaz/nyampur/data/remote/HttpClient.android.kt
@@ -1,4 +1,4 @@
-package id.nisyafawwaz.nyampur
+package id.nisyafawwaz.nyampur.data.remote
 
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.okhttp.OkHttp
