+++ shared/src/commonMain/kotlin/id/nisyafawwaz/nyampur/di/AppModules.kt
@@ -1,3 +1,8 @@
 package id.nisyafawwaz.nyampur.di
 
-val appModules = listOf(networkModule, dataSourceModule, repositoryModule, useCaseModules)
+val appModules = listOf(
+    networkModule,
+    dataSourceModule,
+    repositoryModule,
+    useCaseModules
+)
