+++ shared/src/commonMain/kotlin/id/nisyafawwaz/nyampur/di/DataSourceModule.kt
@@ -1,6 +1,6 @@
 package id.nisyafawwaz.nyampur.di
 
-import id.nisyafawwaz.nyampur.data.remote.datasources.RemoteDataSource
+import id.nisyafawwaz.nyampur.data.remote.datasources.api.RemoteDataSource
 import id.nisyafawwaz.nyampur.data.remote.datasources.supabase.SupabaseAuthDataSource
 import id.nisyafawwaz.nyampur.data.remote.datasources.supabase.SupabaseRecipeDataSource
 import org.koin.core.module.dsl.singleOf
