+++ shared/src/commonMain/kotlin/id/nisyafawwaz/nyampur/di/NetworkModules.kt
@@ -1,7 +1,7 @@
 package id.nisyafawwaz.nyampur.di
 
 import id.nisyafawwaz.nyampur.data.remote.SupabaseConfig
-import id.nisyafawwaz.nyampur.httpClient
+import id.nisyafawwaz.nyampur.data.remote.httpClient
 import org.koin.dsl.module
 
 val networkModule = module {
